+++ translations/de.php
@@ -1,10 +1,10 @@
 <?php
 
 return [
-    'grommasdietz.kirby-locale.label' => 'Sprache',
-    'grommasdietz.kirby-locale.dialog.headline' => 'Sprache für Auswahl setzen',
-    'grommasdietz.kirby-locale.dialog.select.label' => 'Sprache',
-    'grommasdietz.kirby-locale.dialog.empty' => 'Keine Sprache',
-    'grommasdietz.kirby-locale.dialog.group.site' => 'Website-Sprachen',
-    'grommasdietz.kirby-locale.dialog.group.other' => 'Weitere Sprachen',
+    'grommasdietz.locale.label' => 'Sprache',
+    'grommasdietz.locale.dialog.headline' => 'Sprache für Auswahl setzen',
+    'grommasdietz.locale.dialog.select.label' => 'Sprache',
+    'grommasdietz.locale.dialog.empty' => 'Keine Sprache',
+    'grommasdietz.locale.dialog.group.site' => 'Website-Sprachen',
+    'grommasdietz.locale.dialog.group.other' => 'Weitere Sprachen',
 ];