+++ translations/es.php
@@ -1,10 +1,10 @@
 <?php
 
 return [
-    'grommasdietz.kirby-locale.label' => 'Idioma',
-    'grommasdietz.kirby-locale.dialog.headline' => 'Elegir idioma para la selección',
-    'grommasdietz.kirby-locale.dialog.select.label' => 'Idioma',
-    'grommasdietz.kirby-locale.dialog.empty' => 'Sin idioma',
-    'grommasdietz.kirby-locale.dialog.group.site' => 'Idiomas del sitio',
-    'grommasdietz.kirby-locale.dialog.group.other' => 'Otros idiomas',
+    'grommasdietz.locale.label' => 'Idioma',
+    'grommasdietz.locale.dialog.headline' => 'Elegir idioma para la selección',
+    'grommasdietz.locale.dialog.select.label' => 'Idioma',
+    'grommasdietz.locale.dialog.empty' => 'Sin idioma',
+    'grommasdietz.locale.dialog.group.site' => 'Idiomas del sitio',
+    'grommasdietz.locale.dialog.group.other' => 'Otros idiomas',
 ];