+++ translations/fr.php
@@ -1,10 +1,10 @@
 <?php
 
 return [
-    'grommasdietz.kirby-locale.label' => 'Langue',
-    'grommasdietz.kirby-locale.dialog.headline' => 'Choisir la langue pour la sélection',
-    'grommasdietz.kirby-locale.dialog.select.label' => 'Langue',
-    'grommasdietz.kirby-locale.dialog.empty' => 'Aucune langue',
-    'grommasdietz.kirby-locale.dialog.group.site' => 'Langues du site',
-    'grommasdietz.kirby-locale.dialog.group.other' => 'Autres langues',
+    'grommasdietz.locale.label' => 'Langue',
+    'grommasdietz.locale.dialog.headline' => 'Choisir la langue pour la sélection',
+    'grommasdietz.locale.dialog.select.label' => 'Langue',
+    'grommasdietz.locale.dialog.empty' => 'Aucune langue',
+    'grommasdietz.locale.dialog.group.site' => 'Langues du site',
+    'grommasdietz.locale.dialog.group.other' => 'Autres langues',
 ];