+++ translations/nb.php
@@ -1,10 +1,10 @@
 <?php
 
 return [
-    'grommasdietz.kirby-locale.label' => 'Språk',
-    'grommasdietz.kirby-locale.dialog.headline' => 'Velg språk for markeringen',
-    'grommasdietz.kirby-locale.dialog.select.label' => 'Språk',
-    'grommasdietz.kirby-locale.dialog.empty' => 'Ingen språkvalg',
-    'grommasdietz.kirby-locale.dialog.group.site' => 'Nettstedets språk',
-    'grommasdietz.kirby-locale.dialog.group.other' => 'Andre språk',
+    'grommasdietz.locale.label' => 'Språk',
+    'grommasdietz.locale.dialog.headline' => 'Velg språk for markeringen',
+    'grommasdietz.locale.dialog.select.label' => 'Språk',
+    'grommasdietz.locale.dialog.empty' => 'Ingen språkvalg',
+    'grommasdietz.locale.dialog.group.site' => 'Nettstedets språk',
+    'grommasdietz.locale.dialog.group.other' => 'Andre språk',
 ];