+++ translations/nl.php
@@ -1,10 +1,10 @@
 <?php
 
 return [
-    'grommasdietz.kirby-locale.label' => 'Taal',
-    'grommasdietz.kirby-locale.dialog.headline' => 'Kies taal voor de selectie',
-    'grommasdietz.kirby-locale.dialog.select.label' => 'Taal',
-    'grommasdietz.kirby-locale.dialog.empty' => 'Geen taal',
-    'grommasdietz.kirby-locale.dialog.group.site' => 'Website-talen',
-    'grommasdietz.kirby-locale.dialog.group.other' => 'Andere talen',
+    'grommasdietz.locale.label' => 'Taal',
+    'grommasdietz.locale.dialog.headline' => 'Kies taal voor de selectie',
+    'grommasdietz.locale.dialog.select.label' => 'Taal',
+    'grommasdietz.locale.dialog.empty' => 'Geen taal',
+    'grommasdietz.locale.dialog.group.site' => 'Website-talen',
+    'grommasdietz.locale.dialog.group.other' => 'Andere talen',
 ];